feat: implement core identity and authentication architecture including database schemas, JWT services, and signup flow

+++ src/PowerBase.Application/Common/Interfaces/ISystemRoleRepository.cs
@@ -2,5 +2,5 @@ namespace PowerBase.Application.Common.Interfaces;
 
 public interface ISystemRoleRepository
 {
-    Task<long> GetIdByCodeAsync(string code, CancellationToken ct = default);
+    Task<int> GetIdByCodeAsync(string code, CancellationToken ct = default);
 }
